requirements.txt
filelock~=3.16.1
pytest~=8.3.3
sympy~=1.13.3
mpmath~=1.3.0
torch~=2.4.1
setuptools~=75.1.0
packaging~=24.1
colorama~=0.4.6
fsspec~=2024.9.0
markupsafe~=3.0.1
pluggy~=1.5.0
exceptiongroup~=1.2.2
iniconfig~=2.0.0
networkx~=3.3
tokenizer~=3.4.5
tomli~=2.0.2